gerenciador/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
packaging==23.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.4
whitenoise==6.5.0
